+++ server/src/main/java/org/allaymc/server/block/component/BlockWaterBaseComponentImpl.java
@@ -128,7 +128,7 @@ public class BlockWaterBaseComponentImpl extends BlockLiquidBaseComponentImpl {
             return false;
         }
 
-        dimension.setBlockState(flownIntoBy.pos(), hardenedBlockState);
+        dimension.setBlockState(flownIntoBy.pos(), event.getHardenedBlockState());
         dimension.addLevelSoundEvent(MathUtils.center(flownIntoBy.pos()), SoundEvent.FIZZ);
         return true;
     }